requirements.txt
asgiref==3.6.0
Django==4.1.7
django-treebeard==4.6.1
djangorestframework==3.14.0
mysqlclient==2.1.1
protobuf==3.20.3
pytz==2022.7.1
sqlparse==0.4.3
